namespace api in v2, refactor spec helpers
