+++ spec/support/request_spec_helper.rb
@@ -1,4 +1,6 @@
 module RequestSpecHelper
+  include ApiSpecHelper
+
   def get_json(url)
     get(url)
     JSON.parse(response.body)
@@ -21,13 +23,5 @@ module RequestSpecHelper
     integration_session.__send__(:process, :options, *args).tap do
       copy_session_variables!
     end
-  end
-
-  def mock_api_doorkeeper_pass
-    user = create(:user, client_slug: "test")
-    dummy_app = double owner: user
-    dummy_token = double :acceptable? => true, application: dummy_app
-    allow_any_instance_of(Api::BaseController).to receive(:doorkeeper_token) {dummy_token}
-    user
   end
 end
